# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: IBooking API
  description: 自习室座位预约系统(IBooking)调用端接口
  contact:
    name: Penglei Mao
    email: penglei.x.mao@gmail.com
  version: 1.0.0
servers:
  - url: http://10.177.44.64/ibooking/
  - url: http://10.177.44.64/ibooking-dev/api/
paths:
...
  /signIns:
    get:
      tags:
        - SignIn
      summary: Get an array of all signIns
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'  
    post:
      tags:
        - SignIn
      summary: Create a new signIn info
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignIn'          
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'          
  /signIns/{signIn_id}:
    get:
      tags:
        - SignIn
      summary: Get the info of the specified signIn
      parameters:
        - name: signIn_id
          in: path
          required: true
          schema:
            type: integer
            format: int64            
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    patch:
      tags:
        - SignIn
      summary: Update the info of the specified signIn
      parameters:
        - name: signIn_id
          in: path
          required: true
          schema:
            type: integer
            format: int64            
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignIn'
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    delete:
      tags:
        - SignIn
      summary: Delete the info of the specified signIn
      parameters:
        - name: signIn_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /signIns/{signIn_id}/reservation:
    get:
      tags:
        - SignIn
      summary: Get the reservation of the specified signIn
      parameters:
        - name: signIn_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'                                  
components:
  schemas:
...
    400:
      type: object
      description: Bad Request
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: 请求格式错误，服务器无法理解该请求
...
    500:
      type: object
      description:  Internal Server Error      
      properties:
        code:
          type: integer
          example: 500
        message:
          type: string
          example: 服务器内部错误
...
    Campus:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: 江湾校区
    Building:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: 智华楼
        alias:
          type: string
          example: JB
        campus:
          $ref: '#/components/schemas/Campus'          
    Room:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: '101'
        building:
          $ref: '#/components/schemas/Building'
        total_capacity:
          type: integer
          format: int64
          example: 50
        power_socket_capacity:
          type: integer
          format: int64
          example: 20
        status:
          type: string
          enum:
            - Open
            - Closed
            - Under maintenance
    Seat:
      type: object
      properties:
        id:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
          example: 1
        power_supply:
          type: boolean
          example: true
        room:
          $ref: '#/components/schemas/Room'
    Reservation:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/User'
        seat:
          $ref: '#/components/schemas/Seat'
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - Confirmed
            - Cancelled
            - Pending
        status_change_time:
          type: string
          format: date-time
    SignIn:
      type: object
      properties:
        id:
          type: integer
          format: int64
        reservation:
          $ref: '#/components/schemas/Reservation'
        status:
          type: string
          enum:
            - Pending
            - SignIn
            - SignOut
            - TimeOut
        sign_in_start_time:
          type: string
          format: date-time
        sign_in_time:
          type: string
          format: date-time
        sign_out_time:
          type: string
          format: date-time
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 001
        name:
          type: string
          example: penglei
        email:
          type: string
          example: penglei.x.mao@gmail.com
        password:
          type: string
          example: 123456
...
  responses:
...
    400:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/400'
...
    500:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/500'            